requirements.txt
speechrecognition
yt-dlp
prompt
pydub
pyaudio
google-generativeai
flask
sounddevice